+++ tools/release-plugins/info-string/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './info-string';
+export * from './manyStrategy'