+++ tools/release-plugins/info-string/lib/manyStrategy/default.ts
@@ -1,0 +1,6 @@
+export function defaultStrategy(v: any) {
+  if (typeof v === 'object') {
+    return JSON.stringify(v);
+  }
+  return '' + v;
+}