+++ tools/release-plugins/info-string/lib/manyStrategy/strategyKeys.ts
@@ -1,0 +1,7 @@
+export const strategyNames = {
+  default: 'default',
+  deleteScope: 'deleteScope',
+  'YYYY-MM-DD-HH-mm-ss': 'YYYY-MM-DD-HH-mm-ss',
+  'YYYY-MM-DD': 'YYYY-MM-DD',
+  'HH-mm-ss': 'HH-mm-ss',
+} as const;